commit after 2014 release

+++ Version1_01/NEXTA-GUI/PlayerSeekBar.h
@@ -53,7 +53,7 @@ public:
    void GetRange(long& start, long& stop);
    void SetRange(long start, long stop);
    long GetPos(), GetPosReal();
-   void SetPos(long pos);
+   void SetPos(float pos);
 
 
    // Overrides
